+++ Changes.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+- add redis enabled setting
+
 ### [1.1.0] - 2022-11-22
 
 - fix: use this.redis_ping during runtime, #26